WIP split streams
